+++ rmk/src/input_device/pmw33xx.rs
@@ -1,4 +1,4 @@
-// PMW3360 Mouse Sensor Driver
+// PMW3360/ PMW3389 Mouse Sensor Driver
 //
 // Ported from kot149s PMW3610 driver:
 // https://github.com/kot149/pmw3610-rs
@@ -11,6 +11,40 @@ use embedded_hal_async::spi::SpiBus;
 
 use crate::input_device::pointing::{InitState, MotionData, PointingDevice, PointingDriver, PointingDriverError};
 
+// ============================================================================
+// Burst register offsets
+// ============================================================================
+const BURST_MOTION_FLAGS: usize = 0;
+#[allow(dead_code)]
+const BURST_OBSERVATION: usize = 1;
+const BURST_DELTA_X_L: usize = 2;
+const BURST_DELTA_X_H: usize = 3;
+const BURST_DELTA_Y_L: usize = 4;
+const BURST_DELTA_Y_H: usize = 5;
+const BURST_DATA_LEN: usize = 6;
+
+// ============================================================================
+// Constants
+// ============================================================================
+const SPI_WRITE: u8 = 0x80; // BIT(7)
+const MOTION_STATUS_MOTION: u8 = 0x80; // BIT(7)
+const MOTION_STATUS_LIFTED: u8 = 0x08; // BIT(4)
+const POWER_UP_RESET_VAL: u8 = 0x5a;
+
+// Timing constants
+const RESET_DELAY_MS: u64 = 50;
+
+// SPI timing constants (from PMW3360 datasheet)
+const T_NCS_SCLK_US: u64 = 1;
+const T_SRAD_US: u64 = 160;
+const T_SRAD_MOTBR_US: u64 = 35;
+const T_SRX_US: u64 = 20 - T_NCS_SCLK_US;
+const T_SWX_US: u64 = 180 - T_SCLK_NCS_WR_US;
+const T_SCLK_NCS_WR_US: u64 = 35 - T_NCS_SCLK_US;
+const T_BEXIT_US: u64 = 1;
+const T_BRSEP_US: u64 = 15;
+
+// PMW3360/ 3389 registers are almost alike. This Enum contains a few extra registers that are not available for the 3360
 #[allow(dead_code)]
 #[derive(Eq, PartialEq, Debug)]
 enum Register {
@@ -28,7 +62,8 @@ enum Register {
     ShutterLower,
     ShutterUpper,
     Control,
-    Config1,
+    Config1L,
+    Config1H,
     Config2,
     AngleTune,
     FrameCapture,
@@ -49,7 +84,9 @@ enum Register {
     SromId,
     MinSqRun,
     RawDataThreshold,
-    Config5,
+    Control2,
+    Config5L,
+    Config5H,
     PowerUpReset,
     Shutdown,
     InverseProductId,
@@ -63,6 +100,10 @@ enum Register {
     LiftConfig,
     RawDataBurst,
     LiftCutoffTune2,
+    LiftCutoffCalTimeout,
+    LiftCutoffCalMinLen,
+    PWMPeriodCnt,
+    PwmWidthCnt,
 }
 
 impl Register {
@@ -82,7 +123,8 @@ impl Register {
             Register::ShutterLower => 0x0b,
             Register::ShutterUpper => 0x0c,
             Register::Control => 0x0d,
-            Register::Config1 => 0x0f,
+            Register::Config1L => 0x0e,
+            Register::Config1H => 0x0f,
             Register::Config2 => 0x10,
             Register::AngleTune => 0x11,
             Register::FrameCapture => 0x12,
@@ -103,7 +145,9 @@ impl Register {
             Register::SromId => 0x2a,
             Register::MinSqRun => 0x2b,
             Register::RawDataThreshold => 0x2c,
-            Register::Config5 => 0x2f,
+            Register::Control2 => 0x2d,
+            Register::Config5L => 0x2f,
+            Register::Config5H => 0x2f,
             Register::PowerUpReset => 0x3a,
             Register::Shutdown => 0x3b,
             Register::InverseProductId => 0x3f,
@@ -117,56 +161,101 @@ impl Register {
             Register::LiftConfig => 0x63,
             Register::RawDataBurst => 0x64,
             Register::LiftCutoffTune2 => 0x65,
+            Register::LiftCutoffCalTimeout => 0x71,
+            Register::LiftCutoffCalMinLen => 0x72,
+            Register::PWMPeriodCnt => 0x73,
+            Register::PwmWidthCnt => 0x74,
         }
     }
 }
 
-// ============================================================================
-// Burst register offsets
-// ============================================================================
-const BURST_MOTION_FLAGS: usize = 0;
-#[allow(dead_code)]
-const BURST_OBSERVATION: usize = 1;
-const BURST_DELTA_X_L: usize = 2;
-const BURST_DELTA_X_H: usize = 3;
-const BURST_DELTA_Y_L: usize = 4;
-const BURST_DELTA_Y_H: usize = 5;
-const BURST_DATA_LEN: usize = 6;
+pub trait Pmw33xxSpec {
+    // sensor type (60 or 89)
+    const TYPENAME: &str;
+    const PRODUCT_ID: u8;
+    // firmware signature
+    const FW_SIG_PID: u8;
+    const FW_SIG_INV_PID: u8;
+    // resolution constants
+    const RES_MIN: u16;
+    const RES_MAX: u16;
+    const RES_STEP: u16;
 
-// ============================================================================
-// Constants
-// ============================================================================
-const PRODUCT_ID_PMW3360: u8 = 0x42;
-const SPI_WRITE: u8 = 0x80; // BIT(7)
-const MOTION_STATUS_MOTION: u8 = 0x80; // BIT(7)
-const MOTION_STATUS_LIFTED: u8 = 0x08; // BIT(4)
-const POWER_UP_RESET_VAL: u8 = 0x5a;
+    /// Write resolution register
+    async fn write_resolution<SPI, CS, MOTION, SPEC>(
+        driver: &mut Pmw33xx<'_, SPI, CS, MOTION, SPEC>,
+        cpi: u16,
+    ) -> Result<(), Pmw33xxError>
+    where
+        SPI: SpiBus,
+        CS: OutputPin,
+        MOTION: InputPin,
+        SPEC: Pmw33xxSpec;
+}
 
-// Timing constants
-const RESET_DELAY_MS: u64 = 50;
+pub struct Pmw3360Spec;
 
-// SPI timing constants (from PMW3360 datasheet)
-const T_NCS_SCLK_US: u64 = 1;
-const T_SRAD_US: u64 = 160;
-const T_SRAD_MOTBR_US: u64 = 35;
-const T_SRX_US: u64 = 20 - T_NCS_SCLK_US;
-const T_SWX_US: u64 = 180 - T_SCLK_NCS_WR_US;
-const T_SCLK_NCS_WR_US: u64 = 35 - T_NCS_SCLK_US;
-const T_BEXIT_US: u64 = 1;
-const T_BRSEP_US: u64 = 15;
+impl Pmw33xxSpec for Pmw3360Spec {
+    const TYPENAME: &str = "60";
+    const PRODUCT_ID: u8 = 0x42;
+    const FW_SIG_PID: u8 = Self::PRODUCT_ID;
+    const FW_SIG_INV_PID: u8 = 0xBD;
+    const RES_MIN: u16 = 100;
+    const RES_MAX: u16 = 12000;
+    const RES_STEP: u16 = 100;
 
-// Resolution constants
-const RES_STEP: u16 = 100;
-const RES_MIN: u16 = 100;
-const RES_MAX: u16 = 12000;
+    async fn write_resolution<SPI, CS, MOTION, SPEC>(
+        driver: &mut Pmw33xx<'_, SPI, CS, MOTION, SPEC>,
+        cpi: u16,
+    ) -> Result<(), Pmw33xxError>
+    where
+        SPI: SpiBus,
+        CS: OutputPin,
+        MOTION: InputPin,
+        SPEC: Pmw33xxSpec,
+    {
+        driver.write_reg(Register::Config1H, ((cpi / Self::RES_STEP) - 1) as u8).await?;
+        Ok(())
+    }
+}
 
-// Firware signature
-const FW_SIG_PID: u8 = PRODUCT_ID_PMW3360;
-const FW_SIG_INV_PID: u8 = 0xBD;
+pub struct Pmw3389Spec;
 
-/// PMW3360 configuration
+impl Pmw33xxSpec for Pmw3389Spec {
+    const TYPENAME: &str = "89";
+    const PRODUCT_ID: u8 = 0x47;
+    const FW_SIG_PID: u8 = Self::PRODUCT_ID;
+    const FW_SIG_INV_PID: u8 = 0xB8;
+    const RES_MIN: u16 = 50;
+    const RES_MAX: u16 = 16000;
+    const RES_STEP: u16 = 50;
+
+    async fn write_resolution<SPI, CS, MOTION, SPEC>(
+        driver: &mut Pmw33xx<'_, SPI, CS, MOTION, SPEC>,
+        cpi: u16,
+    ) -> Result<(), Pmw33xxError>
+    where
+        SPI: SpiBus,
+        CS: OutputPin,
+        MOTION: InputPin,
+        SPEC: Pmw33xxSpec,
+    {
+        if !(Self::RES_MIN..=Self::RES_MAX).contains(&cpi) {
+            return Err(Pmw33xxError::InvalidCpi);
+        }
+
+        let cpi = (cpi / Self::RES_STEP) - 1;
+        driver.write_reg(Register::Config1H, (cpi >> 8) as u8).await?;
+        driver.write_reg(Register::Config1L, (cpi & 0xFF) as u8).await?;
+
+        Ok(())
+    }
+}
+
+
+/// PMW33xx configuration
 #[derive(Clone)]
-pub struct Pmw3360Config {
+pub struct Pmw33xxConfig {
     /// CPI resolution (100-12000, step 100)
     pub res_cpi: u16,
     /// rot_trans_angle (-127 to 127
@@ -181,7 +270,7 @@ pub struct Pmw3360Config {
     pub swap_xy: bool,
 }
 
-impl Default for Pmw3360Config {
+impl Default for Pmw33xxConfig {
     fn default() -> Self {
         Self {
             res_cpi: 1600,
@@ -194,10 +283,10 @@ impl Default for Pmw3360Config {
     }
 }
 
-/// PMW3360 error types
+/// PMW33xx error types
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub enum Pmw3360Error {
+pub enum Pmw33xxError {
     /// SPI communication error
     Spi,
     /// Invalid product ID detected
@@ -210,29 +299,32 @@ pub enum Pmw3360Error {
     InvalidFwSignature((u8, u8)),
 }
 
-/// PMW3360 driver using embedded-hal SPI traits
-pub struct Pmw3360<'a, SPI, CS, MOTION>
+/// PMW33xx driver using embedded-hal SPI traits
+pub struct Pmw33xx<'a, SPI, CS, MOTION, SPEC>
 where
     SPI: SpiBus,
     CS: OutputPin,
     MOTION: InputPin,
+    SPEC: Pmw33xxSpec,
 {
     spi: SPI,
     cs: CS,
     motion_gpio: Option<MOTION>,
-    config: Pmw3360Config,
+    config: Pmw33xxConfig,
     in_burst: bool,
     srom_firmware: Option<&'a [u8]>,
+    _spec: core::marker::PhantomData<SPEC>,
 }
 
-impl<'a, SPI, CS, MOTION> Pmw3360<'a, SPI, CS, MOTION>
+impl<'a, SPI, CS, MOTION, SPEC> Pmw33xx<'a, SPI, CS, MOTION, SPEC>
 where
     SPI: SpiBus,
     CS: OutputPin,
     MOTION: InputPin,
+    SPEC: Pmw33xxSpec,
 {
-    /// Create a new PMW3360 driver instance
-    pub fn new(spi: SPI, cs: CS, motion_gpio: Option<MOTION>, config: Pmw3360Config) -> Self {
+    /// Create a new PMW33xx driver instance
+    pub fn new(spi: SPI, cs: CS, motion_gpio: Option<MOTION>, config: Pmw33xxConfig) -> Self {
         Self {
             spi,
             cs,
@@ -240,15 +332,16 @@ where
             config,
             in_burst: false,
             srom_firmware: None,
+            _spec: core::marker::PhantomData,
         }
     }
 
-    /// Create a new PMW3360 driver instance with firmware (SROM)
+    /// Create a new PMW33xx driver instance with firmware (SROM)
     pub fn new_with_firmware(
         spi: SPI,
         cs: CS,
         motion_gpio: Option<MOTION>,
-        config: Pmw3360Config,
+        config: Pmw33xxConfig,
         firmware: &'a [u8],
     ) -> Self {
         Self {
@@ -258,52 +351,53 @@ where
             config,
             in_burst: false,
             srom_firmware: Some(firmware),
+            _spec: core::marker::PhantomData,
         }
     }
 
     /// Self check firmware signature of the sensor
-    pub async fn check_fw_signature(&mut self) -> Result<(), Pmw3360Error> {
+    pub async fn check_fw_signature(&mut self) -> Result<(), Pmw33xxError> {
         let product_id = self.read_reg(Register::ProductId).await?;
         let inverse_product_id = self.read_reg(Register::InverseProductId).await?;
 
-        if product_id == FW_SIG_PID && inverse_product_id == FW_SIG_INV_PID {
+        if product_id == SPEC::FW_SIG_PID && inverse_product_id == SPEC::FW_SIG_INV_PID {
             Ok(())
         } else {
             error!(
-                "Firmware signature check failed, expected: {}, {} got: {}, {}",
-                FW_SIG_PID, FW_SIG_INV_PID, product_id, inverse_product_id
+                "PMW33{}: Firmware signature check failed, expected: {}, {} got: {}, {}",
+                SPEC::TYPENAME, SPEC::FW_SIG_PID, SPEC::FW_SIG_INV_PID, product_id, inverse_product_id
             );
-            Err(Pmw3360Error::InvalidFwSignature((product_id, inverse_product_id)))
+            Err(Pmw33xxError::InvalidFwSignature((product_id, inverse_product_id)))
         }
     }
 
     /// Set sensor resolution in CPI (100-12000, step 100)
-    pub async fn set_resolution(&mut self, cpi: u16) -> Result<(), Pmw3360Error> {
-        if !(RES_MIN..=RES_MAX).contains(&cpi) {
-            return Err(Pmw3360Error::InvalidCpi);
+    pub async fn set_resolution(&mut self, cpi: u16) -> Result<(), Pmw33xxError> {
+        if !(SPEC::RES_MIN..=SPEC::RES_MAX).contains(&cpi) {
+            return Err(Pmw33xxError::InvalidCpi);
         }
 
-        self.write_reg(Register::Config1, ((cpi / RES_STEP) - 1) as u8).await?;
+        SPEC::write_resolution(self, cpi).await?;
 
-        debug!("PMW3360: Resolution set to {} CPI", cpi);
+        debug!("PMW33{}: Resolution set to {} CPI", SPEC::TYPENAME, cpi);
 
         Ok(())
     }
 
     /// Set sensor rotational transform angle (-127 to 127)
-    pub async fn set_rot_trans_angle(&mut self, angle: i8) -> Result<(), Pmw3360Error> {
+    pub async fn set_rot_trans_angle(&mut self, angle: i8) -> Result<(), Pmw33xxError> {
         self.write_reg(Register::AngleTune, angle as u8).await?;
 
-        debug!("PMW3360: Rotational transform angle set to {}", angle);
+        debug!("PMW33{}: Rotational transform angle set to {}",SPEC::TYPENAME, angle);
 
         Ok(())
     }
 
     /// Set sensor liftoff distance
-    pub async fn set_liftoff_dist(&mut self, dist: u8) -> Result<(), Pmw3360Error> {
+    pub async fn set_liftoff_dist(&mut self, dist: u8) -> Result<(), Pmw33xxError> {
         self.write_reg(Register::LiftConfig, dist).await?;
 
-        debug!("PMW3360: Liftoff distance set to {}", dist);
+        debug!("PMW33{}: Liftoff distance set to {}", SPEC::TYPENAME, dist);
 
         Ok(())
     }
@@ -315,7 +409,7 @@ where
         }
     }
 
-    async fn read_reg(&mut self, register: Register) -> Result<u8, Pmw3360Error> {
+    async fn read_reg(&mut self, register: Register) -> Result<u8, Pmw33xxError> {
         let _ = self.cs.set_low();
         Timer::after(Duration::from_micros(T_NCS_SCLK_US)).await;
 
@@ -323,12 +417,12 @@ where
         self.spi
             .write(&[register.value() & 0x7f])
             .await
-            .map_err(|_| Pmw3360Error::Spi)?;
+            .map_err(|_| Pmw33xxError::Spi)?;
 
         Timer::after(Duration::from_micros(T_SRAD_US)).await;
 
         let mut value = [0u8];
-        self.spi.read(&mut value).await.map_err(|_| Pmw3360Error::Spi)?;
+        self.spi.read(&mut value).await.map_err(|_| Pmw33xxError::Spi)?;
 
         Self::short_delay();
         let _ = self.cs.set_high();
@@ -338,7 +432,7 @@ where
         Ok(value[0])
     }
 
-    async fn read_burst(&mut self, register: Register, data: &mut [u8]) -> Result<(), Pmw3360Error> {
+    async fn read_burst(&mut self, register: Register, data: &mut [u8]) -> Result<(), Pmw33xxError> {
         let _ = self.cs.set_low();
         Timer::after(Duration::from_micros(T_NCS_SCLK_US)).await;
 
@@ -346,11 +440,11 @@ where
         self.spi
             .write(&[register.value() & 0x7f])
             .await
-            .map_err(|_| Pmw3360Error::Spi)?;
+            .map_err(|_| Pmw33xxError::Spi)?;
 
         Timer::after(Duration::from_micros(T_SRAD_MOTBR_US)).await;
 
-        self.spi.read(data).await.map_err(|_| Pmw3360Error::Spi)?;
+        self.spi.read(data).await.map_err(|_| Pmw33xxError::Spi)?;
 
         Self::short_delay();
         let _ = self.cs.set_high();
@@ -360,7 +454,7 @@ where
         Ok(())
     }
 
-    async fn write_reg(&mut self, register: Register, value: u8) -> Result<(), Pmw3360Error> {
+    async fn write_reg(&mut self, register: Register, value: u8) -> Result<(), Pmw33xxError> {
         let _ = self.cs.set_low();
         Timer::after(Duration::from_micros(T_NCS_SCLK_US)).await;
 
@@ -368,7 +462,7 @@ where
         self.spi
             .write(&[register.value() | SPI_WRITE, value])
             .await
-            .map_err(|_| Pmw3360Error::Spi)?;
+            .map_err(|_| Pmw33xxError::Spi)?;
 
         Timer::after(Duration::from_micros(T_SCLK_NCS_WR_US)).await;
         let _ = self.cs.set_high();
@@ -378,18 +472,18 @@ where
         Ok(())
     }
 
-    async fn configure(&mut self) -> Result<(), Pmw3360Error> {
+    async fn configure(&mut self) -> Result<(), Pmw33xxError> {
         // Power-up reset
         self.write_reg(Register::PowerUpReset, POWER_UP_RESET_VAL).await?;
         Timer::after(Duration::from_millis(RESET_DELAY_MS)).await;
 
         // Verify product ID
         let val = self.read_reg(Register::ProductId).await?;
-        if val != PRODUCT_ID_PMW3360 {
-            error!("Invalid product id: {:#02x}", val);
-            return Err(Pmw3360Error::InvalidProductId(val));
+        if val != SPEC::PRODUCT_ID {
+            error!("PWM33{}: Invalid product id: {:#02}", SPEC::TYPENAME, val);
+            return Err(Pmw33xxError::InvalidProductId(val));
         }
-        info!("PMW3360 detected, product ID: {:#02x}", val);
+        info!("PMW33{} detected, product ID: {:#02x}", SPEC::TYPENAME, val);
 
         // Power-up init sequence
         // Read motion registers to clear them
@@ -410,15 +504,15 @@ where
 
         self.check_fw_signature().await?;
 
-        info!("PMW3360 initialized successfully");
+        info!("PMW33{} initialized successfully", SPEC::TYPENAME);
         Ok(())
     }
 
-    async fn upload_firmware(&mut self, firmware: &[u8]) -> Result<(), Pmw3360Error> {
+    async fn upload_firmware(&mut self, firmware: &[u8]) -> Result<(), Pmw33xxError> {
         self.write_reg(Register::Config2, 0x00).await?; // disable REST mode
 
         let srom_id = firmware[1];
-        info!("PMW3360: Uploading SROM firmware with SROM-Id 0x{:02x}", srom_id);
+        info!("PMW33{}: Uploading SROM firmware with SROM-Id 0x{:02x}", SPEC::TYPENAME, srom_id);
 
         self.write_reg(Register::SromEnable, 0x1d).await?;
         Timer::after(Duration::from_millis(10)).await;
@@ -430,13 +524,13 @@ where
         self.spi
             .write(&[Register::SromLoadBurst.value() | SPI_WRITE])
             .await
-            .map_err(|_| Pmw3360Error::Spi)?;
+            .map_err(|_| Pmw33xxError::Spi)?;
 
         Timer::after(Duration::from_micros(T_SCLK_NCS_WR_US)).await;
 
         for &byte in firmware {
-            debug!("PMW3360: Uploading srom byte: 0x{:02x}", byte);
-            self.spi.write(&[byte]).await.map_err(|_| Pmw3360Error::Spi)?;
+            debug!("PMW33{}: Uploading srom byte: 0x{:02x}", SPEC::TYPENAME, byte);
+            self.spi.write(&[byte]).await.map_err(|_| Pmw33xxError::Spi)?;
             Timer::after(Duration::from_micros(T_BRSEP_US)).await;
         }
 
@@ -447,11 +541,11 @@ where
         let flashed_srom_id = self.read_reg(Register::SromId).await?;
         if srom_id != flashed_srom_id {
             error!(
-                "PMW3360: SROM Firmware upload failed, expected SROM-Id 0x{:02x}, but got 0x{:02x} from the sensor.",
+                "PMW33{}: SROM Firmware upload failed, expected SROM-Id 0x{:02x}, but got 0x{:02x} from the sensor.", SPEC::TYPENAME,
                 srom_id, flashed_srom_id
             );
         } else {
-            info!("PMW3360: Upload successfull, new SROM-Id: 0x{:02x}", flashed_srom_id);
+            info!("PMW33{}: Upload successfull, new SROM-Id: 0x{:02x}", SPEC::TYPENAME, flashed_srom_id);
         }
 
         self.write_reg(Register::Config2, 0x00).await?;
@@ -460,11 +554,12 @@ where
     }
 }
 
-impl<'a, SPI, CS, MOTION> PointingDriver for Pmw3360<'a, SPI, CS, MOTION>
+impl<'a, SPI, CS, MOTION, SPEC> PointingDriver for Pmw33xx<'a, SPI, CS, MOTION, SPEC>
 where
     SPI: SpiBus,
     CS: OutputPin,
     MOTION: InputPin,
+    SPEC: Pmw33xxSpec,
 {
     /// Initialize the sensor (public API)
     async fn init(&mut self) -> Result<(), PointingDriverError> {
@@ -489,11 +584,11 @@ where
             .await
             .map_err(|_| PointingDriverError::Spi)?;
 
-        debug!("PMW3360: Burst raw data {:?}", burst_data);
+        debug!("PMW33{}: Burst raw data {:?}", SPEC::TYPENAME, burst_data);
 
         // panic recovery, sometimes burst mode works weird.
         if (burst_data[BURST_MOTION_FLAGS] & 0b111) != 0x00 {
-            debug!("PMW3360: Burst panic recovery");
+            debug!("PMW33{}: Burst panic recovery", SPEC::TYPENAME);
             self.in_burst = false;
         }
 
@@ -517,7 +612,7 @@ where
             (dx, dy) = (dy, dx);
         }
 
-        debug!("PMW3360 motion: x: {}, y: {}", dx, dy);
+        debug!("PMW33{} motion: x: {}, y: {}", SPEC::TYPENAME, dx, dy);
 
         Ok(MotionData { dx, dy })
     }
@@ -531,69 +626,71 @@ where
     }
 }
 
-/// PMW3360 as an InputDevice for RMK
-impl<'a, SPI, CS, MOTION> PointingDevice<Pmw3360<'a, SPI, CS, MOTION>>
+/// PMW33xx as an InputDevice for RMK
+impl<'a, SPI, CS, MOTION, SPEC> PointingDevice<Pmw33xx<'a, SPI, CS, MOTION, SPEC>>
 where
     SPI: SpiBus,
     CS: OutputPin,
     MOTION: InputPin,
+    SPEC: Pmw33xxSpec,
 {
-    /// Create a new PMW3360 device
-    pub fn new(spi: SPI, cs: CS, motion_gpio: Option<MOTION>, config: Pmw3360Config) -> Self {
+    /// Create a new PMW33xx device
+    pub fn new(spi: SPI, cs: CS, motion_gpio: Option<MOTION>, config: Pmw33xxConfig) -> Self {
         Self {
-            sensor: Pmw3360::new(spi, cs, motion_gpio, config),
+            sensor: Pmw33xx::new(spi, cs, motion_gpio, config),
             init_state: InitState::Pending,
             poll_interval: Duration::from_micros(500),
         }
     }
 
-    /// Create a new PMW3360 device with custom poll interval
+    /// Create a new PMW33xx device with custom poll interval
     pub fn with_poll_interval(
         spi: SPI,
         cs: CS,
         motion_gpio: Option<MOTION>,
-        config: Pmw3360Config,
+        config: Pmw33xxConfig,
         poll_interval_us: u64,
     ) -> Self {
         Self {
-            sensor: Pmw3360::new(spi, cs, motion_gpio, config),
+            sensor: Pmw33xx::new(spi, cs, motion_gpio, config),
             init_state: InitState::Pending,
             poll_interval: Duration::from_micros(poll_interval_us),
         }
     }
 
-    /// Create a new PWM3360 device with SROM firmware
+    /// Create a new PMW33xx device with SROM firmware
     ///
     /// Firmware is downloaded to the sensor on every startup
     pub fn new_with_firmware(
         spi: SPI,
         cs: CS,
         motion_gpio: Option<MOTION>,
-        config: Pmw3360Config,
+        config: Pmw33xxConfig,
         firmware: &'a [u8],
     ) -> Self {
         Self {
-            sensor: Pmw3360::new_with_firmware(spi, cs, motion_gpio, config, firmware),
+            sensor: Pmw33xx::new_with_firmware(spi, cs, motion_gpio, config, firmware),
             init_state: InitState::Pending,
             poll_interval: Duration::from_micros(500),
         }
     }
 
-    /// Create a new PWM3360 device with SROM firmware and custom poll intervall
+    /// Create a new PMW33xx device with SROM firmware and custom poll intervall
     ///
     /// Firmware is downloaded to the sensor on every startup
     pub fn new_with_firmware_poll_interval(
         spi: SPI,
         cs: CS,
         motion_gpio: Option<MOTION>,
-        config: Pmw3360Config,
+        config: Pmw33xxConfig,
         poll_interval_us: u64,
         firmware: &'a [u8],
     ) -> Self {
         Self {
-            sensor: Pmw3360::new_with_firmware(spi, cs, motion_gpio, config, firmware),
+            sensor: Pmw33xx::new_with_firmware(spi, cs, motion_gpio, config, firmware),
             init_state: InitState::Pending,
             poll_interval: Duration::from_micros(poll_interval_us),
         }
     }
 }
+